Java project to store informations of students'
This project is a web portal designed for a college which is used to store information including the personal details, academic performance as well as extracurricular activities in a database; so that it can easily be accessed. It needs a web application to show the progress graph of any student.
It should save the student’s progress level and allow the teachers as well as students to see their progress graph and improve themselves as required.
In general, this project aims to enhance efficiency and at the same time maintain information accurateness.